# backend/audit-service/docs/swagger.yaml
definitions:
...
  models.AuditMandateResponse:
    properties:
      created_at:
        type: string
      effective_date:
        type: string
      expiry_date:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      legal_basis:
        type: string
      mandate_source:
        type: string
      reference_number:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateAuditMandateRequest:
    properties:
      effective_date:
        type: string
      expiry_date:
        type: string
      legal_basis:
        type: string
      mandate_source:
        maxLength: 100
        type: string
      reference_number:
        maxLength: 100
        type: string
      title:
        maxLength: 200
        type: string
    required:
    - effective_date
    - reference_number
    - title
    type: object
...
  models.UpdateAuditMandateRequest:
    properties:
      effective_date:
        type: string
      expiry_date:
        type: string
      is_active:
        type: boolean
      legal_basis:
        type: string
      mandate_source:
        maxLength: 100
        type: string
      title:
        maxLength: 200
        type: string
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorDetail'
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /api/v1/audit-mandates:
    get:
      description: Get a list of audit mandates with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      - description: Active status filter
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List Audit Mandates
      tags:
      - audit-mandates
    post:
      consumes:
      - application/json
      description: Create a new audit mandate
      parameters:
      - description: Create audit mandate request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAuditMandateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create Audit Mandate
      tags:
      - audit-mandates
  /api/v1/audit-mandates/{id}:
    delete:
      description: Delete an audit mandate
      parameters:
      - description: Audit Mandate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete Audit Mandate
      tags:
      - audit-mandates
    get:
      description: Get an audit mandate by ID
      parameters:
      - description: Audit Mandate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuditMandateResponse'
              type: object
      security:
      - Bearer: []
      summary: Get Audit Mandate
      tags:
      - audit-mandates
    put:
      consumes:
      - application/json
      description: Update an audit mandate
      parameters:
      - description: Audit Mandate ID
        in: path
        name: id
        required: true
        type: string
      - description: Update audit mandate request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAuditMandateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update Audit Mandate
      tags:
      - audit-mandates
  /api/v1/audit-mandates/active:
    get:
      description: Get the active audit mandate
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuditMandateResponse'
              type: object
      security:
      - Bearer: []
      summary: Get Active Audit Mandate
      tags:
      - audit-mandates
  /api/v1/audit-mandates/reference:
    get:
      description: Get an audit mandate by reference number
      parameters:
      - description: Reference Number
        in: query
        name: reference
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuditMandateResponse'
              type: object
      security:
      - Bearer: []
      summary: Get Audit Mandate By Reference
      tags:
      - audit-mandates
swagger: "2.0"